requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
dateparser==1.2.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
prettytable==3.12.0
propcache==0.2.0
pycryptodome==3.21.0
python-binance==1.0.22
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.11.6
requests==2.32.3
six==1.16.0
termcolor==2.5.0
typing_extensions==4.12.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
wcwidth==0.2.13
websockets==14.0
yarl==1.17.1
